requirements.txt
# Core deep learning frameworks
torch>=1.10.0
torchvision>=0.11.0

# Vision Transformer models (for CTransPath)
timm>=0.6.0

# Data processing and scientific computing
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=0.24.0

# Image processing
Pillow>=8.0.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Configuration and utilities
PyYAML>=5.4.0
tqdm>=4.50.0

# Optional: For RAR file extraction (if needed)
# rarfile>=4.0

# Optional: For better performance
# opencv-python>=4.5.0

# Note: If using conda, some packages might already be installed
# You can check with: conda list

